upgrade create workspace ui ux
