+++ grouping_flutter/lib/space/presentation/view_models/create_workspace_view_model.dart
@@ -1,4 +1,6 @@
 import 'package:flutter/material.dart';
+import 'package:grouping_project/app/presentation/providers/message_service.dart';
+import 'package:grouping_project/core/shared/message_entity.dart';
 import 'package:grouping_project/core/theme/color.dart';
 import 'package:grouping_project/space/data/datasources/local_data_source/workspace_local_data_source.dart';
 import 'package:grouping_project/space/data/datasources/remote_data_source/workspace_remote_data_source.dart';
@@ -13,27 +15,32 @@ import 'package:image_picker/image_picker.dart';
 
 class CreateWorkspaceViewModel extends ChangeNotifier {
 
+  final MessageService messageService = MessageService();
+  // user data provider for token & current user
   UserDataProvider? userDataProvider;
+  // new workspace data to create
   WorkspaceEntity newWorkspaceData = WorkspaceEntity.newWorkspace();
+  // temp avatar file before upload to server
   XFile? tempAvatarFile;
-
+  // tag for new workspace
   String tag = "";
-  
+  // space colors
   List<Color> spaceColors = AppColor.spaceColors;
-
+  // space color index
   Color get spaceColor => AppColor.getWorkspaceColorByIndex(newWorkspaceData.themeColor);
+  // create state
 
   void clearTag(){
     tag = "";
     notifyListeners();
   }
+  
   void deleteTag(int index) {
     newWorkspaceData.tags.removeAt(index);
     notifyListeners();
   }
 
-  void addTag() {
-    // newWorkspaceData.tags.add(WorkspaceTagModel(content: tag));
+  void addTag(String value ) {
     if(tag.isNotEmpty){
       newWorkspaceData.tags.add(WorkspaceTagModel(content: tag));
       clearTag();
@@ -61,7 +68,22 @@ class CreateWorkspaceViewModel extends ChangeNotifier {
     notifyListeners();
   }
 
-  Future<void> createWorkspace() async {
+  String? spaceNameValidator (String? value) {
+    if(value == null || value.isEmpty){
+      return "小組名稱請勿留空";
+    }
+    return null;
+  }
+
+  String? spaceDescriptionValidator (String? value) {
+    if(value == null || value.isEmpty){
+      return "小組介紹請勿留空";
+    }
+    return null;
+  }
+
+  
+  Future<bool> createWorkspace() async {
 
     debugPrint(newWorkspaceData.toString());
     var workspaceRepo = WorkspaceRepositoryImpl(
@@ -81,7 +103,13 @@ class CreateWorkspaceViewModel extends ChangeNotifier {
 
     workspaceOrFailure.fold(
       (failure) => {
-        debugPrint('create workspace failure: $failure')
+        debugPrint('create workspace failure: $failure'),
+          messageService.addMessage(MessageData.error(
+            title: "新增小組失敗",
+            message: "$failure",
+          ),
+          autoClear: false
+        ),
       },
       (workspace) => {
         newWorkspaceData = workspace,
@@ -91,7 +119,7 @@ class CreateWorkspaceViewModel extends ChangeNotifier {
     );
     
     if(!isSuccess){
-      return;
+      return false;
     }
 
     var joinOrFailure = await joinWorkspaceUseCase(newWorkspaceData, userDataProvider!.currentUser!);
@@ -106,6 +134,8 @@ class CreateWorkspaceViewModel extends ChangeNotifier {
 
     await userDataProvider!.updateUser();
     notifyListeners();
+
+    return true;
   }
 
   void update(UserDataProvider userDataProvider) {
